memcached: Initial cut
Mempool is a preallocated list of 1 MB pages. These pages are
used by the slab allocators which cuts these pages into chunks
of desired sizes and adds them to the slab's free list.

IImplements a barebones slab allocator. This is still a work
in progress at this point, but there are around 10 slabs and the item
chunk size in the next consecuitive samples have a growth rate of 1.25.
The back end pages to the slab caches can be added dynamically as and
when the slabs keep running out of item chunks. Implement a slab cache
containing an Item store for the memcached operation. The slab cache
gets a preallocate page at init time and then on the backend pages are
allocated on demand, i.e. when the slab runs out of items.

This patch adds code to implement timestamps for KV pairs. Chose the
CLOCK_MONOTONIC clock to record timestamps. We do not need a high
precision timestamp, we only need to know if an entry in the cache
expired with reasonable accuracy. The user is expected to send us the
optional timestamp and the user *must* ensure that the expiry timestamp
is accurate if using gettimeofday(). The Cache eviction rules are added
as code comments.